refactor(tests): reorganize into unit/ and integration/ directories
Move all existing test files into tests/unit/. Extract the Plugin+EventHandler
integration tests that were embedded in test_response.py into a dedicated
tests/integration/test_plugin_methods.py, keeping test_response.py focused on
handle_response unit tests only.
